Decompose Composer monolith into focused modules
Replace the 673-line Composer.tsx with 10 focused files under
app/components/composer/:

- Composer.tsx: Shell (layout, dialog state, composition)
- ComposerTextarea.tsx: Textarea element + event wiring
- ComposerFooter.tsx: Single portal decision point (inline vs mobile dock)
- ComposerToolbar.tsx: Action button row
- ComposerErrorBanner.tsx: Auto-dismissing error display
- useComposerActions.ts: Clear, speak, fix-text, enter, live typing
- useComposerTabs.ts: Tab-text synchronization
- useTextareaScroll.ts: Scroll snapshot/intent system
- types.ts: Shared types
- index.ts: Re-export

Key win: ComposerFooter eliminates the dual render path — one component
tree that portals itself on mobile instead of duplicating every footer
element in conditional blocks.

All features preserved: multi-tab, undo/clear, double-enter, live
typing, fix text, tone control, reply suggestions, save as phrase.

Closes #531
