+++ app/components/composer/ComposerErrorBanner.tsx
@@ -1,0 +1,13 @@
+'use client';
+
+interface ComposerErrorBannerProps {
+  error: string;
+}
+
+export default function ComposerErrorBanner({ error }: ComposerErrorBannerProps) {
+  return (
+    <div className="px-6 pb-2">
+      <span className="text-sm text-red-400">{error}</span>
+    </div>
+  );
+}
