+++ app/components/composer/index.ts
@@ -1,0 +1,2 @@
+export { default } from './Composer';
+export type { ComposerProps } from './types';
